Added Template Scene and Fixed some movement
Created very basic template which others can add their mechanics to. Made changes to movement so that after grappling player will maintain upward momentum. Also made it so that wall jump will only be half as strong upward when continuing previous upward momentum so that upward momentum can be conserved in a way that isn't broken.