Some bugfixes and change to grapple logic
Made it so that tension is always mantained in the grapple, not only when velocity.y < 0. Ensured if player backtracked to another room then immediately returned they wouldn't respawn in the former room. Ensured only 1 grapple every time the ground is left. Fixed an issue with ice patches and the y velocity.