Timer Fix + Example
Fixed the timer resetting on death. How the timer now functions is as follows: There is a variable in the GM that stores the current time. The only way the time is reset is when the restart level button is pressed. Time spent per stage includes retries due to death.