Moving Platform, 1 Way, and Grapple Only
Added moving platforms that subtract their velocity from the player when the player lands on them and adds it afterwards to give a sense of gained momentum from the platform. Added a platform that can only be grappled to and collided with on 1 side. Made a grapple-only object that won't give the player their grapple or jumps back. (I realize this is a pretty big commit, I should've committed after creating the moving platforms, and again after creating the 1 Ways)